+++ azure-storage-blob/ChangeLog.md
@@ -4,6 +4,7 @@
 
 ## Version XX.XX.XX:
 - Fixed design flaw where get_blob_to_* methods buffer entire blob when max_connections is set to 1.
+- Added support for access conditions on append_blob_from_* methods.
 
 ## Version 1.3.0: